requirements_prod.txt
amqp==5.1.0
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.3.2
cryptography==36.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.3
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-debug-toolbar-force==0.1.8
django-environ==0.8.1
django-filter==21.1
django-nine==0.2.5
django-test==0.4030
djangorestframework==3.13.1
flake8==4.0.1
flower==1.0.0
hackernote==0.0.2
humanize==4.0.0
idna==3.3
insultgenerator==1.0.3
kombu==5.2.4
loremipsum==1.0.5
mccabe==0.6.1
oauthlib==3.2.0
packaging==21.3
probed==0.0.9
prometheus-client==0.14.0
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.7
python-crontab==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
redis==4.2.2
requests==2.27.1
requests-oauthlib==1.3.0
shared==0.0.13
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tornado==6.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
White-Noise==0.1.0
wrapt==1.14.0
